requirements.txt
mediapipe==0.9.2.1
number-recognition==1.0.5
opencv-python==4.7.0.72
pyyaml==6.0